Handle digits that are in names.
There are scenarios where the names and/or other details in the first element from the MRZ code array are represented by digits instead of characters. For exampe, P<UTO3RIKSS0N<<ANNA<MARIA<<<<<<<<<<<<<<<<<<< instead of P<UTOERIKSSON<<ANNA<MARIA<<<<<<<<<<<<<<<<<<<.

This fix serves to handle these kind of scenarios so that digits are mapped to there correct/right equivalent Characters so that we do not have issues when applying regex checks on the results.